Migrated to SDL 3

+++ src/abstract/ComponentSDL.cpp
@@ -2,7 +2,7 @@
 // File       : ComponentSDL.cpp
 // Author     : riyufuchi
 // Created on : Feb 21, 2025
-// Last edit  : Feb 21, 2025
+// Last edit  : Dec 02, 2025
 // Copyright  : Copyright (c) 2025, riyufuchi
 // Description: ConsoleArt
 //==============================================================================
@@ -12,11 +12,11 @@
 namespace sdl
 {
 
-ComponentSDL::ComponentSDL(int x, int y, int width, int height, std::function<void()> onClickEvent) : ComponentSDL(x, y, width, height, {128, 128, 128, 255}, onClickEvent)
+ComponentSDL::ComponentSDL(float x, float y, float width, float height, std::function<void()> onClickEvent) : ComponentSDL(x, y, width, height, {128, 128, 128, 255}, onClickEvent)
 {
 }
 
-ComponentSDL::ComponentSDL(int x, int y, int width, int height, SDL_Color baseColor, std::function<void()>  onClickEvent) : mouseOver(false),baseColor(baseColor), rect({x, y, width, height}), onClickEvent(onClickEvent)
+ComponentSDL::ComponentSDL(float x, float y, float width, float height, SDL_Color baseColor, std::function<void()>  onClickEvent) : mouseOver(false),baseColor(baseColor), rect({x, y, width, height}), onClickEvent(onClickEvent)
 {
 }
 
@@ -24,21 +24,21 @@ ComponentSDL::~ComponentSDL()
 {
 }
 
-void ComponentSDL::repose(int x, int y)
+void ComponentSDL::repose(float x, float y)
 {
 	rect.x = x;
 	rect.y = y;
 }
 
-void ComponentSDL::center(int width, int height)
+void ComponentSDL::center(float width, float height)
 {
-	rect.x = std::max(0, (width - rect.w) / 2);
-	rect.y = std::max(0, (height - rect.h) / 2);
+	rect.x = std::max(0.0f, (width - rect.w) / 2);
+	rect.y = std::max(0.0f, (height - rect.h) / 2);
 }
 
-void ComponentSDL::centerHorizontal(int width)
+void ComponentSDL::centerHorizontal(float width)
 {
-	rect.x = std::max(0, (width - rect.w) / 2);
+	rect.x = std::max(0.0f, (width - rect.w) / 2);
 }
 
 void ComponentSDL::onClick()
@@ -47,27 +47,27 @@ void ComponentSDL::onClick()
 		onClickEvent();
 }
 
-bool ComponentSDL::isMouseInside(int mouseX, int mouseY) const
+bool ComponentSDL::isMouseInside(float mouseX, float mouseY) const
 {
 	return (mouseX >= rect.x && mouseX <= rect.x + rect.w && mouseY >= rect.y && mouseY <= rect.y + rect.h);
 }
 
-void ComponentSDL::setX(int x)
+void ComponentSDL::setX(float x)
 {
 	this->rect.x = x;
 }
 
-void ComponentSDL::setY(int y)
+void ComponentSDL::setY(float y)
 {
 	this->rect.y = y;
 }
 
-void ComponentSDL::setWidth(int width)
+void ComponentSDL::setWidth(float width)
 {
 	this->rect.w = width;
 }
 
-void ComponentSDL::setHeight(int height)
+void ComponentSDL::setHeight(float height)
 {
 	this->rect.h = height;
 }
@@ -77,22 +77,22 @@ void ComponentSDL::setMouseOver(bool mouseOver)
 	this->mouseOver = mouseOver;
 }
 
-int ComponentSDL::getX() const
+float ComponentSDL::getX() const
 {
 	return this->rect.x;
 }
 
-int ComponentSDL::getY() const
+float ComponentSDL::getY() const
 {
 	return this->rect.y;
 }
 
-int ComponentSDL::getWidth() const
+float ComponentSDL::getWidth() const
 {
 	return this->rect.w;
 }
 
-int ComponentSDL::getHeight() const
+float ComponentSDL::getHeight() const
 {
 	return this->rect.h;
 }
